fix typos (ninetyninepercentcasino!)
