+++ core/bin/main/MainCasino.gwt.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "https://www.gwtproject.org/doctype/2.10.0/gwt-module.dtd">
 <module>
-	<source path="com/ninetyninenercentcasino/game" />
+	<source path="com/ninetyninepercentcasino/game" />
 </module>
