+++ core/src/com/ninetyninepercentcasino/game/PasswordIncorrect.java
@@ -1,4 +1,4 @@
-package com.ninetyninenercentcasino.game;
+package com.ninetyninepercentcasino.game;
 
 public class PasswordIncorrect extends Exception{
     public PasswordIncorrect() {
