+++ core/src/com/ninetyninepercentcasino/game/UserAlreadyExists.java
@@ -1,4 +1,4 @@
-package com.ninetyninenercentcasino.game;
+package com.ninetyninepercentcasino.game;
 
 public class UserAlreadyExists extends Exception {
     public UserAlreadyExists() {
