+++ desktop/build.gradle
@@ -2,7 +2,7 @@ sourceCompatibility = 1.8
 sourceSets.main.java.srcDirs = [ "src/" ]
 sourceSets.main.resources.srcDirs = ["../assets"]
 
-project.ext.mainClassName = "com.ninetyninenercentcasino.game.DesktopLauncher"
+project.ext.mainClassName = "com.ninetyninepercentcasino.game.DesktopLauncher"
 project.ext.assetsDir = new File("../assets")
 
 import org.gradle.internal.os.OperatingSystem
