--- html/src/com/ninetyninenercentcasino/game/GdxDefinition.gwt.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "https://www.gwtproject.org/doctype/2.10.0/gwt-module.dtd">
-<module rename-to="html">
-	<inherits name='com.badlogic.gdx.backends.gdx_backends_gwt' />
-	<inherits name='com.badlogic.gdx.physics.box2d.box2d-gwt' />
-	<inherits name='Box2DLights' />
-
-	<inherits name='MainCasino' />
-	<entry-point class='com.ninetyninenercentcasino.game.client.HtmlLauncher' />
-	<set-configuration-property name='xsiframe.failIfScriptTag' value='FALSE'/>
-	<set-configuration-property name="gdx.assetpath" value="../assets" />
-	<set-property name="user.agent" value="gecko1_8, safari"/>
-	<collapse-property name="user.agent" values="*" />
-</module>
