--- html/src/com/ninetyninenercentcasino/game/GdxDefinitionSuperdev.gwt.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "https://www.gwtproject.org/doctype/2.10.0/gwt-module.dtd">
-<module rename-to="html">
-	<inherits name='com.badlogic.gdx.backends.gdx_backends_gwt' />
-	<inherits name='com.badlogic.gdx.physics.box2d.box2d-gwt' />
-	<inherits name='Box2DLights' />
-
-    <inherits name='com.ninetyninenercentcasino.game.GdxDefinition' />
-
-    <collapse-all-properties />
-
-	<add-linker name="xsiframe"/>
-	<set-configuration-property name="devModeRedirectEnabled" value="true"/>
-	<set-configuration-property name='xsiframe.failIfScriptTag' value='FALSE'/>
-</module>
