--- html/src/com/ninetyninenercentcasino/game/client/HtmlLauncher.java
@@ -1,22 +1,0 @@
-package com.ninetyninenercentcasino.game.client;
-
-import com.badlogic.gdx.ApplicationListener;
-import com.badlogic.gdx.backends.gwt.GwtApplication;
-import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
-import com.ninetyninenercentcasino.game.MainCasino;
-
-public class HtmlLauncher extends GwtApplication {
-
-        @Override
-        public GwtApplicationConfiguration getConfig () {
-                // Resizable application, uses available space in browser
-                return new GwtApplicationConfiguration(true);
-                // Fixed size application:
-                //return new GwtApplicationConfiguration(480, 320);
-        }
-
-        @Override
-        public ApplicationListener createApplicationListener () {
-                return new MainCasino();
-        }
-}
